requirements.txt
# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0

# Machine Learning
xgboost>=1.7.0
pycaret>=3.0.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Deep Learning (Optional)
torch>=2.0.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
pathlib2>=2.3.7

# Image Processing (if needed)
Pillow>=9.0.0
scikit-image>=0.19.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.0.0

# Testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0

